Use uintptr_t for engine data
Gets rid of a warning on 32-bit builds on Linux.

+++ file.h
@@ -97,7 +97,7 @@ struct fio_file {
 	/*
 	 * For use by the io engine
 	 */
-	uint64_t engine_data;
+	uintptr_t engine_data;
 
 	/*
 	 * if io is protected by a semaphore, this is set
